pagination now working

+++ routes/web.php
@@ -4,7 +4,10 @@ use App\Http\Controllers\PagesController;
 use \App\Http\Controllers\ProductsController;
 use \App\Http\Controllers\JobController;
 use App\Http\Controllers\TodoController;
+use \App\Http\Controllers\CustomAuthController;
 use Illuminate\Support\Facades\Route;
+use App\Models\JobPost;
+
 
 /*
 |--------------------------------------------------------------------------
@@ -19,16 +22,27 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('/',[JobController::class, 'index'])->name('index');
 Route::get('/site/about',[JobController::class, 'about'])->name('about');
-Route::get('/site/signup',[JobController::class, 'signup'])->name('signup');
-Route::get('/site/login',[JobController::class, 'login'])->name('login');
+//  Route::get('/site/signup',[JobController::class, 'signup'])->name('signup');
+//  Route::get('/site/login',[JobController::class, 'login'])->name('login');
 Route::get('/job/jobs',[JobController::class, 'jobs'])->name('jobs');
 Route::get('/job/job/{id}',[JobController::class, 'jobDetails'])->name('job');
 
-Route::get('/job/createJob',[JobController::class,'createJob'])->name('createJob');
-Route::post('/job/storeJob',[JobController::class,'storeJob'])->name('storeJob');
-Route::get('/job/editJob/{id}',[JobController::class,'editJob'])->name('editJob');
-Route::put('/job/updateJob/{id}',[JobController::class,'updateJob'])->name('updateJob');
-Route::delete('/job/deleteJob/{id}',[JobController::class,'deleteJob'])->name('deleteJob');
+Route::get('/job/createJob',[JobController::class,'createJob'])->name('createJob')->middleware('auth');;
+Route::post('/job/storeJob',[JobController::class,'storeJob'])->name('storeJob')->middleware('auth');;
+Route::get('/job/editJob/{id}',[JobController::class,'editJob'])->name('editJob')->middleware('auth');;
+Route::put('/job/updateJob/{id}',[JobController::class,'updateJob'])->name('updateJob')->middleware('auth');;
+Route::delete('/job/deleteJob/{id}',[JobController::class,'deleteJob'])->name('deleteJob')->middleware('auth');;
+
+
+
+//These routes are for authentication
+
+Route::get('dashboard', [CustomAuthController::class, 'dashboard']); 
+Route::get('login', [CustomAuthController::class, 'index'])->name('login');
+Route::post('custom-login', [CustomAuthController::class, 'customLogin'])->name('login.custom'); 
+Route::get('registration', [CustomAuthController::class, 'registration'])->name('register-user');
+Route::post('custom-registration', [CustomAuthController::class, 'customRegistration'])->name('register.custom'); 
+Route::get('signout', [CustomAuthController::class, 'signOut'])->name('signout');
